+++ include/openenclave/internal/report.h
@@ -45,7 +45,7 @@ typedef struct _oe_get_quote_args
 {
     oe_result_t result;
     sgx_report_t sgxReport;
-    uint32_t quoteSize;
+    size_t quoteSize;
     uint8_t quote[1];
 } oe_get_quote_args_t;
 
@@ -63,10 +63,10 @@ typedef struct _oe_get_report_args
     uint32_t flags; /* in */
 
     uint8_t optParams[sizeof(sgx_target_info_t)]; /* in */
-    uint32_t optParamsSize;                       /* in */
+    size_t optParamsSize;                         /* in */
 
-    uint8_t* reportBuffer;     /* ptr to output buffer */
-    uint32_t reportBufferSize; /* in-out */
+    uint8_t* reportBuffer;   /* ptr to output buffer */
+    size_t reportBufferSize; /* in-out */
 } oe_get_report_args_t;
 
 /*
@@ -80,8 +80,8 @@ typedef struct _oe_verify_report_args
 {
     oe_result_t result; /* out */
 
-    uint8_t* report;     /* in */
-    uint32_t reportSize; /* in */
+    uint8_t* report;   /* in */
+    size_t reportSize; /* in */
 } oe_verify_report_args_t;
 
 /*
@@ -93,18 +93,18 @@ typedef struct _oe_verify_report_args
 */
 typedef struct _oe_get_revocation_info_args
 {
-    oe_result_t result;                /* out */
-    uint8_t fmspc[6];                  /* in */
-    const char* crl_urls[3];           /* in */
-    uint32_t num_crl_urls;             /* in */
-    uint8_t* tcb_info;                 /* out */
-    uint32_t tcb_info_size;            /* out */
-    uint8_t* tcb_issuer_chain;         /* out */
-    uint32_t tcb_issuer_chain_size;    /* out */
-    uint8_t* crl[3];                   /* out */
-    uint32_t crl_size[3];              /* out */
-    uint8_t* crl_issuer_chain[3];      /* out */
-    uint32_t crl_issuer_chain_size[3]; /* out */
+    oe_result_t result;              /* out */
+    uint8_t fmspc[6];                /* in */
+    const char* crl_urls[3];         /* in */
+    uint32_t num_crl_urls;           /* in */
+    uint8_t* tcb_info;               /* out */
+    size_t tcb_info_size;            /* out */
+    uint8_t* tcb_issuer_chain;       /* out */
+    size_t tcb_issuer_chain_size;    /* out */
+    uint8_t* crl[3];                 /* out */
+    size_t crl_size[3];              /* out */
+    uint8_t* crl_issuer_chain[3];    /* out */
+    size_t crl_issuer_chain_size[3]; /* out */
 
     // Memory allocated by host to pass outputs back to the enclave. Enclave
     // must free this memory via oe_host_free.
